Placed timecap of 10s on dialogue
