+++ Assets/Scripts/Audio/ObjectVoicelineManager.cs
@@ -6,17 +6,16 @@ using UnityEngine.Scripting.APIUpdating;
 using UnityEngine.UI;
 using UnityEngine.EventSystems;
 using UnityEngine.Audio;
+using System;
 
 
-/* This script is a dirty fix, ideally those would be managed by animation
- * timelines.
- */
-public class TVReactionVoicelineManager : MonoBehaviour
+public class ObjectVoicelineManager : MonoBehaviour
 {
-    public static TVReactionVoicelineManager instance;
+    public static ObjectVoicelineManager instance;
 
-    private TapeReactions _tapeReactions = null;
-    private string _voiceline;
+    [Header("Info on Held Item")]
+    private PickupInteractable _pickupInteractable;
+    private string _inspectionObjectText;
 
     [Header("Audio")]
     private AudioClip _dialogueAudio;
@@ -28,8 +27,6 @@ public class TVReactionVoicelineManager : MonoBehaviour
     private Coroutine resetCoroutine;
     private Coroutine scrollCoroutine;
 
-    private TapeManager tapeManager;
-
     private void Awake()
     {
         if (instance == null)
@@ -40,25 +37,31 @@ public class TVReactionVoicelineManager : MonoBehaviour
         {
             Destroy(this);
         }
-
-        tapeManager = FindObjectOfType<TapeManager>();
     }
 
     void Start()
     {
-        _voiceline = "Nothing Yet"; //tODO: REPLACE WITH EMPTY
-        VideoControls.dialoguePrompt += Play;
+        _inspectionObjectText = "Nothing Yet"; //tODO: REPLACE WITH EMPTY
+        PickUpInteractor.OnObjectPlace += HandlePlace;
     }
 
     private void OnDestroy()
     {
-        VideoControls.dialoguePrompt -= Play;
+        PickUpInteractor.OnObjectPlace -= HandlePlace;
     }
+   
+    void HandlePlace(GameObject obj) 
+    {
+        Stop();
 
+        _pickupInteractable = null;
+        _inspectionObjectText = "";
+    }
+    
     public void Play()
     {
-        Debug.Log("Playing Dialogue: " + _voiceline);
-        float timer = _dialogueAudio.length;
+        Debug.Log("Playing Dialogue: " + _inspectionObjectText);
+        float timer = Math.Min(_dialogueAudio.length, 10f);
 
         TMP_Text textField = subtitlesText;
 
@@ -84,48 +87,20 @@ public class TVReactionVoicelineManager : MonoBehaviour
     }
 
     #region Set Dialogue + Subtitles
+    public void SetDialogue(GameObject obj)
+    {
+        Debug.Log("Setting Dialogue for: " + obj.name);
+        _pickupInteractable = obj.GetComponent<PickupInteractable>();
+        _inspectionObjectText = _pickupInteractable.inspectionObjectText;
+        _dialogueAudio = _pickupInteractable.dialogueAudio;
+    }
+
     public void SetDialogue(string text, AudioClip dialogueAudio)
     {
         Debug.Log("Setting Dialogue for: " + text);
-        _voiceline = text;
+        _inspectionObjectText = text;
         _dialogueAudio = dialogueAudio;
         //VideoControls.dialoguePrompt += PlayDialogue;
-    }
-
-    public float ChangeReactionVoicelineForTape(ClipToPlay currentClip)
-    {
-        Debug.Log("Changing voiceline for tape");
-        float timer = 0.0f;
-
-        switch(currentClip)
-        {
-            case ClipToPlay.OriginalCorrupted:
-                SetDialogue(_tapeReactions.startSubtitles, _tapeReactions.start);
-                timer = 0.1f;
-                break;
-
-            case ClipToPlay.BranchACorrupted:
-                SetDialogue(_tapeReactions.middleSubtitles, _tapeReactions.middle);
-                timer = 0.8f;
-                break;
-
-            case ClipToPlay.BranchBCorrupted:
-                SetDialogue(_tapeReactions.middleSubtitles, _tapeReactions.middle);
-                timer = 0.8f;
-                break;
-
-            case ClipToPlay.BranchASolution:
-                SetDialogue(_tapeReactions.endASubtitles, _tapeReactions.endA);
-                timer = 0.8f;
-                break;
-
-            case ClipToPlay.BranchBSolution:
-                SetDialogue(_tapeReactions.endBSubtitles, _tapeReactions.endB);
-                timer = 0.8f;
-                break;
-        }
-
-        return timer;
     }
     #endregion
 
@@ -142,7 +117,7 @@ public class TVReactionVoicelineManager : MonoBehaviour
     {
         Debug.Log("Turning ON subtitles");
 
-        uitext.text = _voiceline;
+        uitext.text = _inspectionObjectText;
         uitext.ForceMeshUpdate();
         uitext.pageToDisplay = 1;
 
@@ -151,7 +126,7 @@ public class TVReactionVoicelineManager : MonoBehaviour
         {
             scrollCoroutine = StartCoroutine(ScrollSubtitles(uitext, timer));
         }
-
+        
         resetCoroutine = StartCoroutine(ResetSubtitles(uitext, timer));
     }
 
@@ -165,7 +140,7 @@ public class TVReactionVoicelineManager : MonoBehaviour
         for (int i = 0; i < pages; i++)
         {
             Debug.Log("Scrolling to page: " + uitext.pageToDisplay);
-
+            
             yield return new WaitForSeconds(timePerPage);
             uitext.pageToDisplay = i + 2;
 
@@ -179,13 +154,4 @@ public class TVReactionVoicelineManager : MonoBehaviour
         Debug.Log("Turning off subtitles");
     }
     #endregion
-
-    private void Update()
-    {
-        if (tapeManager.televisionHasTape())
-        {
-            _tapeReactions = tapeManager.GetTapeReactionsInTV();
-        }
-    }
 }
-
