+++ Assets/Scripts/Audio/TVReactionVoicelineManager.cs
@@ -6,15 +6,18 @@ using UnityEngine.Scripting.APIUpdating;
 using UnityEngine.UI;
 using UnityEngine.EventSystems;
 using UnityEngine.Audio;
+using System;
 
 
-public class ObjectVoicelineManager : MonoBehaviour
+/* This script is a dirty fix, ideally those would be managed by animation
+ * timelines.
+ */
+public class TVReactionVoicelineManager : MonoBehaviour
 {
-    public static ObjectVoicelineManager instance;
+    public static TVReactionVoicelineManager instance;
 
-    [Header("Info on Held Item")]
-    private PickupInteractable _pickupInteractable;
-    private string _inspectionObjectText;
+    private TapeReactions _tapeReactions = null;
+    private string _voiceline;
 
     [Header("Audio")]
     private AudioClip _dialogueAudio;
@@ -26,6 +29,8 @@ public class ObjectVoicelineManager : MonoBehaviour
     private Coroutine resetCoroutine;
     private Coroutine scrollCoroutine;
 
+    private TapeManager tapeManager;
+
     private void Awake()
     {
         if (instance == null)
@@ -36,31 +41,25 @@ public class ObjectVoicelineManager : MonoBehaviour
         {
             Destroy(this);
         }
+
+        tapeManager = FindObjectOfType<TapeManager>();
     }
 
     void Start()
     {
-        _inspectionObjectText = "Nothing Yet"; //tODO: REPLACE WITH EMPTY
-        PickUpInteractor.OnObjectPlace += HandlePlace;
+        _voiceline = "Nothing Yet"; //tODO: REPLACE WITH EMPTY
+        VideoControls.dialoguePrompt += Play;
     }
 
     private void OnDestroy()
     {
-        PickUpInteractor.OnObjectPlace -= HandlePlace;
+        VideoControls.dialoguePrompt -= Play;
     }
-   
-    void HandlePlace(GameObject obj) 
-    {
-        Stop();
 
-        _pickupInteractable = null;
-        _inspectionObjectText = "";
-    }
-    
     public void Play()
     {
-        Debug.Log("Playing Dialogue: " + _inspectionObjectText);
-        float timer = _dialogueAudio.length;
+        Debug.Log("Playing Dialogue: " + _voiceline);
+        float timer = Math.Min(_dialogueAudio.length, 10f);
 
         TMP_Text textField = subtitlesText;
 
@@ -86,20 +85,48 @@ public class ObjectVoicelineManager : MonoBehaviour
     }
 
     #region Set Dialogue + Subtitles
-    public void SetDialogue(GameObject obj)
-    {
-        Debug.Log("Setting Dialogue for: " + obj.name);
-        _pickupInteractable = obj.GetComponent<PickupInteractable>();
-        _inspectionObjectText = _pickupInteractable.inspectionObjectText;
-        _dialogueAudio = _pickupInteractable.dialogueAudio;
-    }
-
     public void SetDialogue(string text, AudioClip dialogueAudio)
     {
         Debug.Log("Setting Dialogue for: " + text);
-        _inspectionObjectText = text;
+        _voiceline = text;
         _dialogueAudio = dialogueAudio;
         //VideoControls.dialoguePrompt += PlayDialogue;
+    }
+
+    public float ChangeReactionVoicelineForTape(ClipToPlay currentClip)
+    {
+        Debug.Log("Changing voiceline for tape");
+        float timer = 0.0f;
+
+        switch(currentClip)
+        {
+            case ClipToPlay.OriginalCorrupted:
+                SetDialogue(_tapeReactions.startSubtitles, _tapeReactions.start);
+                timer = 0.1f;
+                break;
+
+            case ClipToPlay.BranchACorrupted:
+                SetDialogue(_tapeReactions.middleSubtitles, _tapeReactions.middle);
+                timer = 0.8f;
+                break;
+
+            case ClipToPlay.BranchBCorrupted:
+                SetDialogue(_tapeReactions.middleSubtitles, _tapeReactions.middle);
+                timer = 0.8f;
+                break;
+
+            case ClipToPlay.BranchASolution:
+                SetDialogue(_tapeReactions.endASubtitles, _tapeReactions.endA);
+                timer = 0.8f;
+                break;
+
+            case ClipToPlay.BranchBSolution:
+                SetDialogue(_tapeReactions.endBSubtitles, _tapeReactions.endB);
+                timer = 0.8f;
+                break;
+        }
+
+        return timer;
     }
     #endregion
 
@@ -116,7 +143,7 @@ public class ObjectVoicelineManager : MonoBehaviour
     {
         Debug.Log("Turning ON subtitles");
 
-        uitext.text = _inspectionObjectText;
+        uitext.text = _voiceline;
         uitext.ForceMeshUpdate();
         uitext.pageToDisplay = 1;
 
@@ -125,7 +152,7 @@ public class ObjectVoicelineManager : MonoBehaviour
         {
             scrollCoroutine = StartCoroutine(ScrollSubtitles(uitext, timer));
         }
-        
+
         resetCoroutine = StartCoroutine(ResetSubtitles(uitext, timer));
     }
 
@@ -139,7 +166,7 @@ public class ObjectVoicelineManager : MonoBehaviour
         for (int i = 0; i < pages; i++)
         {
             Debug.Log("Scrolling to page: " + uitext.pageToDisplay);
-            
+
             yield return new WaitForSeconds(timePerPage);
             uitext.pageToDisplay = i + 2;
 
@@ -153,4 +180,13 @@ public class ObjectVoicelineManager : MonoBehaviour
         Debug.Log("Turning off subtitles");
     }
     #endregion
+
+    private void Update()
+    {
+        if (tapeManager.televisionHasTape())
+        {
+            _tapeReactions = tapeManager.GetTapeReactionsInTV();
+        }
+    }
 }
+
